[klippa] code reorg
split subsetting code for each lookup type into separate files
Also export Intersect trait as it's needed by subsetting as well
