+++ read-fonts/src/tables/gpos/closure.rs
@@ -8,7 +8,7 @@ use super::{
 use crate::{collections::IntSet, GlyphId, ReadError, Tag};
 
 #[cfg(feature = "std")]
-use crate::tables::layout::{LayoutLookupList, LookupClosure, LookupClosureCtx};
+use crate::tables::layout::{Intersect, LayoutLookupList, LookupClosure, LookupClosureCtx};
 
 impl Gpos<'_> {
     /// Return a set of all feature indices underneath the specified scripts, languages and features
@@ -85,7 +85,9 @@ impl LookupClosure for PositionLookup<'_> {
 
         self.subtables()?.closure_lookups(c, lookup_index)
     }
+}
 
+impl Intersect for PositionLookup<'_> {
     fn intersects(&self, glyph_set: &IntSet<GlyphId>) -> Result<bool, ReadError> {
         self.subtables()?.intersects(glyph_set)
     }
@@ -99,7 +101,9 @@ impl LookupClosure for PositionSubtables<'_> {
             _ => Ok(()),
         }
     }
+}
 
+impl Intersect for PositionSubtables<'_> {
     fn intersects(&self, glyph_set: &IntSet<GlyphId>) -> Result<bool, ReadError> {
         match self {
             PositionSubtables::Single(subtables) => subtables.intersects(glyph_set),
@@ -114,7 +118,7 @@ impl LookupClosure for PositionSubtables<'_> {
     }
 }
 
-impl LookupClosure for SinglePos<'_> {
+impl Intersect for SinglePos<'_> {
     fn intersects(&self, glyph_set: &IntSet<GlyphId>) -> Result<bool, ReadError> {
         match self {
             Self::Format1(item) => item.intersects(glyph_set),
@@ -123,19 +127,19 @@ impl LookupClosure for SinglePos<'_> {
     }
 }
 
-impl LookupClosure for SinglePosFormat1<'_> {
+impl Intersect for SinglePosFormat1<'_> {
     fn intersects(&self, glyph_set: &IntSet<GlyphId>) -> Result<bool, ReadError> {
         Ok(self.coverage()?.intersects(glyph_set))
     }
 }
 
-impl LookupClosure for SinglePosFormat2<'_> {
+impl Intersect for SinglePosFormat2<'_> {
     fn intersects(&self, glyph_set: &IntSet<GlyphId>) -> Result<bool, ReadError> {
         Ok(self.coverage()?.intersects(glyph_set))
     }
 }
 
-impl LookupClosure for PairPos<'_> {
+impl Intersect for PairPos<'_> {
     fn intersects(&self, glyph_set: &IntSet<GlyphId>) -> Result<bool, ReadError> {
         match self {
             Self::Format1(item) => item.intersects(glyph_set),
@@ -144,7 +148,7 @@ impl LookupClosure for PairPos<'_> {
     }
 }
 
-impl LookupClosure for PairPosFormat1<'_> {
+impl Intersect for PairPosFormat1<'_> {
     fn intersects(&self, glyph_set: &IntSet<GlyphId>) -> Result<bool, ReadError> {
         let coverage = self.coverage()?;
         let pair_sets = self.pair_sets();
@@ -176,7 +180,7 @@ impl LookupClosure for PairPosFormat1<'_> {
     }
 }
 
-impl LookupClosure for PairSet<'_> {
+impl Intersect for PairSet<'_> {
     fn intersects(&self, glyph_set: &IntSet<GlyphId>) -> Result<bool, ReadError> {
         for record in self.pair_value_records().iter() {
             let second_glyph = record?.second_glyph();
@@ -188,33 +192,33 @@ impl LookupClosure for PairSet<'_> {
     }
 }
 
-impl LookupClosure for PairPosFormat2<'_> {
+impl Intersect for PairPosFormat2<'_> {
     fn intersects(&self, glyph_set: &IntSet<GlyphId>) -> Result<bool, ReadError> {
         Ok(self.coverage()?.intersects(glyph_set) && self.class_def2()?.intersects(glyph_set)?)
     }
 }
 
-impl LookupClosure for CursivePosFormat1<'_> {
+impl Intersect for CursivePosFormat1<'_> {
     fn intersects(&self, glyph_set: &IntSet<GlyphId>) -> Result<bool, ReadError> {
         Ok(self.coverage()?.intersects(glyph_set))
     }
 }
 
-impl LookupClosure for MarkBasePosFormat1<'_> {
+impl Intersect for MarkBasePosFormat1<'_> {
     fn intersects(&self, glyph_set: &IntSet<GlyphId>) -> Result<bool, ReadError> {
         Ok(self.mark_coverage()?.intersects(glyph_set)
             && self.base_coverage()?.intersects(glyph_set))
     }
 }
 
-impl LookupClosure for MarkLigPosFormat1<'_> {
+impl Intersect for MarkLigPosFormat1<'_> {
     fn intersects(&self, glyph_set: &IntSet<GlyphId>) -> Result<bool, ReadError> {
         Ok(self.mark_coverage()?.intersects(glyph_set)
             && self.ligature_coverage()?.intersects(glyph_set))
     }
 }
 
-impl LookupClosure for MarkMarkPosFormat1<'_> {
+impl Intersect for MarkMarkPosFormat1<'_> {
     fn intersects(&self, glyph_set: &IntSet<GlyphId>) -> Result<bool, ReadError> {
         Ok(self.mark1_coverage()?.intersects(glyph_set)
             && self.mark2_coverage()?.intersects(glyph_set))
